Interface

AudioSchedules.AudioSchedule

AudioSchedules.Cycles

AudioSchedules.Envelope

AudioSchedules.Line

AudioSchedules.Plan

AudioSchedules.StrictMap

AudioSchedules.Synthesizer

AudioSchedules.equal_loudness

AudioSchedules.extrema!

AudioSchedules.make_iterator

AudioSchedules.plan_within

AudioSchedules.schedule!

AudioSchedules.@q_str

AudioSchedules.AudioSchedule — Method

AudioSchedule(the_sample_rate)

Create an AudioSchedule.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> a_schedule = AudioSchedule();

Add a synthesizer to the schedule with schedule!.

julia> envelope = Envelope((0, 0.25, 0), (0.05s, 0.95s), (Line, Line));

julia> schedule!(a_schedule, StrictMap(sin, Cycles(440Hz)), 0s, envelope)

julia> schedule!(a_schedule, StrictMap(sin, Cycles(440Hz)), 1s, envelope)

julia> schedule!(a_schedule, StrictMap(sin, Cycles(550Hz)), 1s, envelope)

Then, you can create a SampledSource from the schedule using Plan.

julia> a_plan = Plan(a_schedule, 44100Hz);

You can find the number of samples in a Plan with length.

julia> the_length = length(a_plan)
88200

You can use the plan as a source for samples.

julia> using SampledSignals: unsafe_read!

julia> buf = Vector{Float64}(undef, the_length);

julia> unsafe_read!(a_plan, buf, 0, the_length);

julia> buf[1:4] ≈ [0.0, 7.1029846e-6, 2.8356127e-5, 6.3592327e-5]
true

AudioSchedules.Cycles — Type

Cycles(frequency)

Cycles from 0 to 2π to repeat at a frequency.

AudioSchedules.Envelope — Type

Envelope(levels, durations, shapes)

Shapes are all functions which return Synthesizers:

shape(start_value, end_value, duration) -> Synthesizer

durations and levels list the time and level of the boundaries of segments of the envelope. For example,

Envelope([0.0, 1.0, 1.0, 0.0], [.05 s, 0.9 s, 0.05 s], [Line, Line, Line])

will create an envelope with three segments:

Line(0.0, 1.0, 0.05 s)
Line(1.0, 1.0, 0.9 s)
Line(1.0, 0.0, 0.05 s)

AudioSchedules.Line — Type

Line(start_value, end_value, duration)

A line from start_value to end_value that lasts for duration.

AudioSchedules.Plan — Method

Plan(a_schedule::AudioSchedule)

Return a SampledSource for the schedule.

AudioSchedules.StrictMap — Type

StrictMap(a_function, synthesizers...)

Map a_function over synthesizers, assuming that none of the synthesizers will end before they are scheduled to.

AudioSchedules.Synthesizer — Type

abstract type Synthesizer

Synthesizers need only support make_iterator.

AudioSchedules.equal_loudness — Method

equal_loudness(synthesizer::StrictMap{<:Any, Tuple{Cycles}})

Change the volume of a synthesizer so that sounds played at different frequencies will have the same perceived volume. Assumes that the map function has a period of 2π.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> soft = equal_loudness(StrictMap(cos, Cycles(10000Hz)));

julia> first(make_iterator(soft, 44100Hz)) ≈ 0.0053035474
true

Technical details: uses the ISO 226:2003 curve for 40 phons. Scales output by a ratio of the equivalent sound pressure at the current frequency to the equivalent sound pressure at 20Hz (about as low as humans can hear).

AudioSchedules.extrema! — Method

extrema!(a_plan::Plan)

Find the extrema of a plan. This will consume the plan.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> a_schedule = AudioSchedule();

julia> schedule!(a_schedule, StrictMap(sin, Cycles(440Hz)), 0s, Envelope((1, 1), (1s,), (Line,)))

julia> extrema!(Plan(a_schedule, 44100Hz)) .≈ (-0.99999974, 0.99999974)
(true, true)

AudioSchedules.make_iterator — Function

make_iterator(synthesizer, the_sample_rate)

Return an iterator that will the play the synthesizer at the_sample_rate

AudioSchedules.plan_within — Method

plan_within(a_schedule::AudioSchedule, the_sample_rate; maximum_volume = 1.0)

Make a plan, then adjust the volume to maximum_volume.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> schedule = AudioSchedule();

julia> note = StrictMap(sin, Cycles(440Hz)), 0s, Envelope((1, 1), (1s,), (Line,));

julia> schedule!(schedule, note...); schedule!(schedule, note...)

julia> plan = plan_within(schedule, 44100Hz);

julia> extrema!(plan) .≈ (-1.0, 1.0)
(true, true)

AudioSchedules.schedule! — Method

schedule!(schedule::AudioSchedule, synthesizer::Synthesizer, start_time, envelope::Envelope)

Schedule an audio synthesizer to be added to the schedule, starting at start_time with the duration and volume contained in an Envelope.

AudioSchedules.@q_str — Macro

q"interval"

Create a musical interval. You can specify a numerator (which defaults to 1) and denominator (which defaults to 1) and an octave shift (which defaults to 0).

julia> using AudioSchedules

julia> q"1"
1//1

julia> q"3/2"
3//2

julia> q"2/3o1"
4//3

julia> q"2/3o-1"
1//3

julia> q"o2"
4//1