Interface

AudioSchedules.AudioSchedule

AudioSchedules.Cycles

AudioSchedules.Envelope

AudioSchedules.Line

AudioSchedules.StrictMap

AudioSchedules.Synthesizer

AudioSchedules.equal_loudness

AudioSchedules.extrema!

AudioSchedules.make_iterator

AudioSchedules.make_plan

AudioSchedules.plan!

AudioSchedules.readjust!

AudioSchedules.schedule!

AudioSchedules.AudioSchedule — Method

AudioSchedule(the_sample_rate)

Create an AudioSchedule.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> a_schedule = AudioSchedule(44100Hz)
AudioSchedule with triggers at () seconds

Add a synthesizer to the schedule with schedule!.

julia> envelope = Envelope((0, 0.25, 0), (0.05s, 0.95s), (Line, Line));

julia> schedule!(a_schedule, StrictMap(sin, Cycles(440Hz)), 0s, envelope)

julia> schedule!(a_schedule, StrictMap(sin, Cycles(440Hz)), 1s, envelope)

julia> schedule!(a_schedule, StrictMap(sin, Cycles(550Hz)), 1s, envelope)

julia> a_schedule
AudioSchedule with triggers at (0.0, 0.05, 1.0, 1.05, 2.0) seconds

Then, you can create a SampledSource from the schedule using plan!.

julia> using SampledSignals: unsafe_read!

julia> a_plan = plan!(a_schedule);

You can find the number of samples in a Plan with length.

julia> the_length = length(a_plan)
88200

You can use the plan as a source for samples.

julia> buf = Vector{Float64}(undef, the_length);

julia> unsafe_read!(a_plan, buf, 0, the_length);

julia> buf[1:4] ≈ [0.0, 7.1029846e-6, 2.8356127e-5, 6.3592327e-5]
true

You can only plan! a schedule once.

julia> plan!(a_schedule)
ERROR: The schedule was empty or had already been consumed

AudioSchedules.Cycles — Type

Cycles(frequency)

Cycles from 0 to 2π to repeat at a frequency.

AudioSchedules.Envelope — Type

Envelope(levels, durations, shapes)

Shapes are all functions which return Synthesizers:

shape(start_value, end_value, duration) -> Synthesizer

durations and levels list the time and level of the boundaries of segments of the envelope. For example,

Envelope([0.0, 1.0, 1.0, 0.0], [.05 s, 0.9 s, 0.05 s], [Line, Line, Line])

will create an envelope with three segments:

Line(0.0, 1.0, 0.05 s)
Line(1.0, 1.0, 0.9 s)
Line(1.0, 0.0, 0.05 s)

AudioSchedules.Line — Type

Line(start_value, end_value, duration)

A line from start_value to end_value that lasts for duration.

AudioSchedules.StrictMap — Type

StrictMap(a_function, synthesizers...)

Map a_function over synthesizers, assuming that none of the synthesizers will end before they are scheduled to.

AudioSchedules.Synthesizer — Type

abstract type Synthesizer

Synthesizers need only support make_iterator.

AudioSchedules.equal_loudness — Method

equal_loudness(synthesizer::StrictMap{<:Any, Tuple{Cycles}})

Change the volume of a synthesizer so that sounds played at different frequencies will have the same perceived volume. Assumes that the map function has a period of 2π. Uses the ISO 226:2003 curves for 40 phons.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> soft = equal_loudness(StrictMap(cos, Cycles(10000Hz)));

julia> first(AudioSchedules.make_iterator(soft, 44100)) ≈ 0.0053035474
true

AudioSchedules.extrema! — Method

extrema!(a_plan::Plan)

Find the extrema of a plan. Unfortunately, this uses up the iterators in a plan.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> a_schedule = AudioSchedule(44100Hz);

julia> schedule!(a_schedule, StrictMap(sin, Cycles(440Hz)), 0s, Envelope((1, 1), (1s,), (Line,)))

julia> extrema!(plan!(a_schedule)) .≈ (-0.99999974, 0.99999974)
(true, true)

AudioSchedules.make_plan — Method

make_plan(the_sample_rate, triples, maximum_volume = 1.0)

A convenience function. triples should be a list of triples in the form

(synthesizer, start_time, envelope)

Create a test schedule with AudioSchedule with the_sample_rate, schedule! all the triples, find the extrema! of the test plan, then repeat all the previous steps, readjust!ing the volume to maximum_volume.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> envelope = Envelope((0, 0.25, 0), (0.05s, 0.95s), (Line, Line));

julia> plan = make_plan(44100Hz, [
            (StrictMap(sin, Cycles(440Hz)), 0s, envelope),
            (StrictMap(sin, Cycles(440Hz)), 1s, envelope),
            (StrictMap(sin, Cycles(550Hz)), 1s, envelope)
        ]);

julia> extrema!(plan) .≈ (-1.0, 0.99895193)
(true, true)

AudioSchedules.plan! — Method

plan!(a_schedule::AudioSchedule)

Return a SampledSource for the schedule.

AudioSchedules.readjust! — Method

readjust!(old_plan::Plan, new_plan::Plan; maximum_volume = 1.0)

Readjust the volume in new_plan based on the volume of old_plan, to a maximum volume of 1.0. Note: will "use up" old_plan.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> schedule_1 = AudioSchedule(44100Hz); schedule_2 = AudioSchedule(44100Hz);

julia> note = (StrictMap(sin, Cycles(440Hz)), 0s, Envelope((1, 1), (1s,), (Line,)));

julia> schedule!(schedule_1, note...); schedule!(schedule_1, note...)

julia> schedule!(schedule_2, note...); schedule!(schedule_2, note...)

julia> plan_2 = plan!(schedule_2);

julia> readjust!(plan!(schedule_1), plan_2)

julia> extrema!(plan_2) .≈ (-1.0, 1.0)
(true, true)

AudioSchedules.schedule! — Method

schedule!(schedule::AudioSchedule, synthesizer::Synthesizer, start_time, envelope::Envelope)

Schedule an audio synthesizer to be added to the schedule, starting at start_time with the duration and volume contained in an Envelope.

AudioSchedules.make_iterator — Function

make_iterator(synthesizer, the_sample_rate)

Return an iterator that will the play the synthesizer at the_sample_rate