Interface

AudioSchedulers.AudioScheduler

AudioSchedulers.Cycles

AudioSchedulers.Envelope

AudioSchedulers.Line

AudioSchedulers.Map

AudioSchedulers.Synthesizer

AudioSchedulers.make_iterator

AudioSchedulers.play

AudioSchedulers.schedule!

AudioSchedulers.AudioScheduler — Method

AudioScheduler(sink)

Create a AudioScheduler to schedule changes to sink.

julia> using AudioSchedulers

julia> using Unitful: s, Hz

julia> using PortAudio: PortAudioStream

julia> stream = PortAudioStream(samplerate = 44100);

julia> scheduler = AudioScheduler(stream.sink)
AudioScheduler with triggers at ()

Add a synthesizer to the schedule with schedule!. You can schedule for a duration in seconds, or use an Envelope.

julia> envelope = Envelope((0, 0.25, 0), (0.05s, 0.95s), (Line, Line));

julia> schedule!(scheduler, Map(sin, Cycles(440Hz)), 0s, envelope)

julia> schedule!(scheduler, Map(sin, Cycles(440Hz)), 1s, envelope)

julia> schedule!(scheduler, Map(sin, Cycles(550Hz)), 1s, envelope)

Then, you can play it with play.

julia> play(scheduler)

You can only play a scheduler once.

julia> play(scheduler)
ERROR: EOFError: read end of file
[...]

julia> close(stream)

AudioSchedulers.Cycles — Type

Cycles(frequency)

Cycles from 0 2π to repeat at a frequency.

AudioSchedulers.Envelope — Type

Envelope(levels, durations, shapes)

Shapes are all functions which return Synthesizers:

shape(start_value, end_value, duration) -> Synthesizer

durations and levels list the time and level of the boundaries of segments of the envelope. For example,

Envelope([0.0, 1.0, 1.0, 0.0], [.05 s, 0.9 s, 0.05 s], [Line, Line, Line])

will create an envelope with three segments:

Line(0.0, 1.0, 0.05 s)
Line(1.0, 1.0, 0.9 s)
Line(s1.0, 0.0, 0.05 s)

See the example for AudioScheduler.

AudioSchedulers.Line — Type

Line(start_value, end_value, duration)

A line from start_value to end_value that lasts for duration.

AudioSchedulers.Map — Type

Map(a_function, synthesizers)

Map a_function over audio synthesizers.

AudioSchedulers.Synthesizer — Type

abstract type Synthesizer

Synthesizers need only support make_iterator.

AudioSchedulers.make_iterator — Method

make_iterator(synthesizer, samplerate)

Return an iterator that will the synthesizer at a given samplerate

AudioSchedulers.play — Method

play(scheduler::AudioScheduler)

Play an AudioScheduler. See the example for AudioScheduler.

AudioSchedulers.schedule! — Method

schedule!(scheduler::AudioScheduler, synthesizer, start_time, duration)

Schedule an audio synthesizer to be added to the scheduler, starting at start_time and lasting for duration. You can also pass an Envelope as a duration. See the example for AudioScheduler. Note: the scheduler will discard the first sample in the iterator during scheduling.