Interface

Performance note

Presumably due to the limits of inference, scheduling 16 or more synthesizers simultaneously will lead you off a performance cliff. Hopefully this limitation will go away in future versions of Julia.

AudioSchedules.AudioSchedule

AudioSchedules.Cycles

AudioSchedules.Grow

AudioSchedules.Hook

AudioSchedules.Line

AudioSchedules.Map

AudioSchedules.SawTooth

AudioSchedules.Scale

AudioSchedules.Weaver

AudioSchedules.add!

AudioSchedules.equal_loudness

AudioSchedules.make_series

AudioSchedules.segments

AudioSchedules.@q_str

AudioSchedules.AudioSchedule — Method

AudioSchedule(; sample_rate = 44100Hz)

Create an empty AudioSchedule. You can add! new synthesizers to the aschedule. Specify a `sampleratewith units per time, like1/sorHz`.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> a_schedule = AudioSchedule()
0.0 s 44100.0 Hz AudioSchedule

julia> add!(
           a_schedule,
           Map(sin, Cycles(440Hz)),
           0s,
           0,
           Line => 1s,
           0.05,
           Line => 1s,
           0.05,
           Line => 1s,
           0,
       )

julia> add!(
            a_schedule,
            Map(sin, Cycles(660Hz)),
            2s,
            0,
            Line => 1s,
            0.05,
            Line => 1s,
            0.05,
            Line => 1s,
            0,
        )

julia> a_schedule
5.0 s 44100.0 Hz AudioSchedule

You can use write the a_schedule directly to a PortAudioStream. However, to do so, the PortAudioStream must have a Weaver writer. The PortAudioStream must have exactly 0 input channels, 1 output channel, and a matching sample rate.

julia> using PortAudio: PortAudioStream

julia> PortAudioStream(0, 1, writer = Weaver()) do stream
           write(stream, a_schedule)
       end

You can save an AudioSchedule as a SampledSignals.SampleBuf.

julia> using SampledSignals: SampleBuf

julia> saved = SampleBuf(a_schedule)
220500-frame, 1-channel SampleBuf{Float64, 2}
5.0s sampled at 44100.0Hz
▂▂▃▃▃▃▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▃▃▃▃▂▂

julia> PortAudioStream(0, 1, warn_xruns = false) do stream
           write(stream, saved)
       end
220500

AudioSchedules.Cycles — Type

Cycles(frequency)

Cycles from 0 to 2π to repeat at a frequency (with frequency units, like Hz). Supports make_series.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> first(make_series(Cycles(440Hz), 44100Hz))
0.06268937721449021

AudioSchedules.Grow — Type

Grow(start_level, rate)

Exponentially grow or decay from start_level (unitless), at a continuous rate (with units per time like 1/s). Supports make_series and segments.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> first(make_series(Grow(1, 1 / s), 44100Hz))
1.0000226759940578

AudioSchedules.Hook — Type

Hook(rate, slope)

Make a hook shape, with an exponential curve growing at a continuous rate (with units per time like 1/s), followed by a line with slope (with units per time like 1/s). Use with add!. Supports segments. Not all hooks are solvable.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> a_schedule = AudioSchedule();

julia> add!(a_schedule, Map(sin, Cycles(440Hz)), 0s, 1, Hook(1 / s, 1 / s) => 2s, ℯ + 1)

julia> add!(a_schedule, Map(sin, Cycles(440Hz)), 0s, 1, Hook(1 / s, 1 / s) => 2s, 0)
ERROR: Unsolvable hook
[...]

AudioSchedules.Line — Type

Line(start_level, slope)

A line from start_level (unitless) with slope (with units per time like 1/s). Supports make_series and segments.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> first(make_series(Line(0, 1 / s), 44100Hz))
2.2675736961451248e-5

AudioSchedules.Map — Type

Map(a_function, synthesizers...)

Map a_function over synthesizers. Supports make_series.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> first(make_series(Map(sin, Cycles(440Hz)), 44100Hz))
0.06264832417874369

AudioSchedules.SawTooth — Method

SawTooth(overtones)

Build a saw-tooth wave from its partials, starting with the fundamental (1), up to overtones.

To increase richness but also buziness, increase overtones.

julia> using AudioSchedules

julia> SawTooth(3)(π / 4)
0.9185207636218614

AudioSchedules.Scale — Type

function Scale(ratio)

A simple wrapper that will multiply inputs by the ratio.

julia> using AudioSchedules

julia> Scale(3)(2)
6

AudioSchedules.Weaver — Type

struct Weaver <: Scribe end

A special PortAudio Scribe. You can use a Weaver to write an AudioSchedule directly to your speakers.

AudioSchedules.add! — Method

add!(a_schedule::AudioSchedule, synthesizer, start_time,
    start_level, shape => duration, end_level, more_segments...
)

Add a synthesizer to a AudioSchedule, where synthesizer is anything that supports make_series, start_time has units of time (like s), and the rest of the arguments specify the shape of the envelope.

For all envelope segments, call

segments(shape, start_level, duration, end_level)

duration should have units of time (like s). For example,

add!(a_schedule, synthesizer, start_time, 0, Line => 1s, 1, Line => 1s, 0)

will call segments twice:

segments(Line, 0, 1s, 1)
segments(Line, 1, 1s, 0)

AudioSchedules.equal_loudness — Method

equal_loudness(synthesizer::Map{<:Any, Tuple{Cycles}})

Change the volume of a synthesizer so that sounds played at different frequencies will have the same perceived volume. Assumes that the map function has a period of 2π.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> soft = equal_loudness(Map(cos, Cycles(10000Hz)));

julia> first(make_series(soft, 44100Hz))
0.0007717680651318246

Technical details: uses the ISO 226:2003 curve for 40 phons. Scales output by a ratio of the equivalent sound pressure at the current frequency to the equivalent sound pressure at 20Hz (about as low as humans can hear).

AudioSchedules.make_series — Method

make_series(synthesizer, sample_rate)

Return an iterator that will the play the synthesizer at sample_rate (with frequency units, like Hz). The iterator should yield Float64s between -1 and 1. Assumes that iterators will never end while they are scheduled.

AudioSchedules.segments — Method

segments(shape, start_level, duration, end_level)

Called for each envelope segment passed to add!. Return a tuple of pairs in the form (segment, duration), where duration has units of time (like s), for a segment of shape shape.

julia> using AudioSchedules

julia> using Unitful: s

julia> segments(Grow, 1, 1s, ℯ)
((Grow(1.0, 1.0 s^-1), 1 s),)

AudioSchedules.@q_str — Macro

q"interval"

Create a musical interval. You can specify a numerator (which defaults to 1), denominator (which defaults to 1), and an octave shift (which defaults to 0).

julia> using AudioSchedules

julia> q"1"
1//1

julia> q"3/2"
3//2

julia> q"2/3o1"
4//3

julia> q"2/3o-1"
1//3

julia> q"o2"
4//1

julia> q"1 + 1"
ERROR: LoadError: Base.Meta.ParseError("Can't parse interval 1 + 1")
[...]