Interface

Performance note

Presumably due to the limits of inference, scheduling 16 or more synthesizers simultaneously will lead you off a performance cliff. Hopefully this limitation will go away in future versions of Julia.

AudioSchedules.AudioSchedule

AudioSchedules.Cycles

AudioSchedules.Grow

AudioSchedules.Hook

AudioSchedules.Line

AudioSchedules.Map

AudioSchedules.Player

AudioSchedules.beats!

AudioSchedules.compound_wave

AudioSchedules.envelope

AudioSchedules.equal_loudness

AudioSchedules.get_duration

AudioSchedules.make_iterator

AudioSchedules.modulate!

AudioSchedules.note!

AudioSchedules.pluck

AudioSchedules.schedule_within

AudioSchedules.segments

AudioSchedules.@q_str

AudioSchedules.AudioSchedule — Method

AudioSchedule(triples, the_sample_rate)

Return a SampledSource. triples should be a vector of triples in the form

(synthesizer, start_time, duration_or_envelope)

where synthesizer is anything that supports make_iterator, start_time has units of time (like s), and duration_or_envelope is either a duration (with units of time, like s) or an envelope.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> an_envelope = envelope(0, Line => 1s, 1, Line => 1s, 0);

julia> triple = (Map(sin, Cycles(440Hz)), 0s, an_envelope);

julia> a_schedule = AudioSchedule([triple], 44100Hz);

You can find the number of samples in an AudioSchedule with length.

julia> the_length = length(a_schedule)
88200

You can use the schedule as a source for samples.

julia> read(a_schedule, the_length);

The schedule must have at least one triple.

julia> AudioSchedule([], 44100Hz)
ERROR: AudioSchedules require at least one triple
[...]

AudioSchedules.Cycles — Type

Cycles(frequency)

Cycles from 0 to 2π to repeat at a frequency (with frequency units, like Hz). Supports make_iterator and segments.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> first(make_iterator(Cycles(440Hz), 44100Hz))
0.0

AudioSchedules.Grow — Type

Grow(start, rate)

Exponentially grow or decay from start (unitless), at a continuous rate (with units per time like 1/s). Supports make_iterator and segments..

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> first(make_iterator(Grow(1, 1 / s), 44100Hz))
1.0

AudioSchedules.Hook — Type

Hook(rate, slope)

Make a hook shape, with an exponential curve growing at a continuous rate (with units per time like 1/s), followed by a line with slope (with units per time like 1/s). Use with envelope. Supports segments.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> envelope(1, Hook(1 / s, 1 / s) => 2s, ℯ + 1)
((Grow(1.0, 1.0 s^-1), 1.0 s), (Line(2.718281828459045, 1.0 s^-1), 1.0 s))

AudioSchedules.Line — Type

Line(start, slope)

A line from start (unitless) with slope (with units per time like 1/s). Supports make_iterator and segments.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> first(make_iterator(Line(0, 1 / s), 44100Hz))
0.0

AudioSchedules.Map — Type

Map(a_function, synthesizers...)

Map a_function over synthesizers. Supports make_iterator.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> first(make_iterator(Map(sin, Cycles(440Hz)), 44100Hz))
0.0

AudioSchedules.Player — Method

Player(triples; wave = compound_wave(Val(7)), make_envelope = pluck, key = 440Hz, clock = 0s, seconds_per_beat = 1s)

An all-included way of composing music. wave should be a a function which takes radians and yields amplitudes between -1 and 1 and defaults to a compound_wave. make_envelope should be a function which takes a duration in units of time (like s) and returns an envelope. It defaults to pluck. key is the current key of the song. clock is the current time. Add notes to the song with note!. Move the clock with beats!. Modulate the key with modulate. When you are finished, you can create an AudioSchedule from the triples.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> player = Player([]);

julia> note!(player, 1, 1)

julia> a_schedule = AudioSchedule(player.triples, 44100Hz);

julia> first(read(a_schedule, length(a_schedule)))
0.0

AudioSchedules.beats! — Method

beats!(player::Player, beats)

Move the Player forward a certain number of beats.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> player = Player([]);

julia> note!(player, 1, 1)

julia> beats!(player, 1)

julia> note!(player, 1, 1)

julia> a_schedule = AudioSchedule(player.triples, 44100Hz);

julia> first(read(a_schedule, length(a_schedule)))
0.0

AudioSchedules.compound_wave — Method

compound_wave(overtones)

Build a saw-tooth wave from its partials, starting with the fundamental (1), up to overtones. You can pass overtones as a integer, or as a Val to maximize performance.

To increase richness but also buziness, increase overtones.

julia> using AudioSchedules

julia> compound_wave(3)(π / 4)
1.4428090415820634

AudioSchedules.envelope — Method

envelope(start_level, shape => duration, end_level, more_segments...)

For all envelope segments, call

segments(start_level, shape, duration, end_level)

duration should have units of time (like s). For example,

envelope(0, Line => 1s, 1, Line => 1s, 0)

will add two segments:

segments(0, Line, 1s, 1)
segments(1, Line, 1s, 0)

julia> using AudioSchedules

julia> using Unitful: s

julia> envelope(0, Line => 1s, 1, Line => 1s, 0)
((Line(0.0, 1.0 s^-1), 1 s), (Line(1.0, -1.0 s^-1), 1 s))

AudioSchedules.equal_loudness — Method

equal_loudness(synthesizer::Map{<:Any, Tuple{Cycles}})

Change the volume of a synthesizer so that sounds played at different frequencies will have the same perceived volume. Assumes that the map function has a period of 2π.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> soft = equal_loudness(Map(cos, Cycles(10000Hz)));

julia> first(make_iterator(soft, 44100Hz)) ≈ 0.0053035474
true

Technical details: uses the ISO 226:2003 curve for 40 phons. Scales output by a ratio of the equivalent sound pressure at the current frequency to the equivalent sound pressure at 20Hz (about as low as humans can hear).

AudioSchedules.get_duration — Method

get_duration(synthesizer)

Get the duration of a synthesizer (with units of time, like s), for synthesizers with an inherent length.

julia> using AudioSchedules

julia> using FileIO: load

julia> import LibSndFile

julia> cd(joinpath(pkgdir(AudioSchedules), "test"))

julia> get_duration(load("clunk.wav"))
0.3518820861678005 s

AudioSchedules.make_iterator — Method

make_iterator(synthesizer, the_sample_rate)

Return an iterator that will the play the synthesizer at the_sample_rate (with frequency units, like Hz). The iterator should yield ratios between -1 and 1. Assumes that iterators will never end while they are scheduled.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> using FileIO: load

julia> import LibSndFile

julia> cd(joinpath(pkgdir(AudioSchedules), "test"))

julia> first(make_iterator(load("clunk.wav"), 44100Hz))
0.00168Q0f15

AudioSchedules.modulate! — Method

modulate!(player::Player, ratio)

Shift the key of the Player by ratio.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> player = Player([]);

julia> note!(player, 1, 1)

julia> modulate!(player, 2)

julia> note!(player, 1, 1)

julia> a_schedule = AudioSchedule(player.triples, 44100Hz);

julia> first(read(a_schedule, length(a_schedule)))
0.0

AudioSchedules.note! — Method

note!(player::Player, ratio, beats)

Add a note to the player. ratio will be the ratio between the current key and the new note, and beats is how many beats the note will play for.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> player = Player([]);

julia> note!(player, 1, 1)

julia> a_schedule = AudioSchedule(player.triples, 44100Hz);

julia> first(read(a_schedule, length(a_schedule)))
0.0

AudioSchedules.pluck — Method

pluck(time; decay = -2.5/s, slope = 1/0.005s, peak = 1)

Make an envelope with an exponential decay (with units per time, like 1/s) from the peak, and ramps with ±slope (in units per time, like 1/s) on each side.

julia> using AudioSchedules

julia> using Unitful: s

julia> pluck(1s)
((Line(0.0, 200.0 s^-1), 0.005 s), (Grow(1.0, -2.5 s^-1), 0.9945839800394016 s), (Line(0.08320399211967063, -200.0 s^-1), 0.0004160199605983683 s))

AudioSchedules.schedule_within — Method

schedule_within(triples, the_sample_rate; maximum_volume = 1.0)

Make an AudioSchedule with triples and the_sample_rate (with frequency units like Hz), then adjust the volume to maximum_volume. Will iterate through triples twice.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> triple = (Map(sin, Cycles(440Hz)), 0s, 1s);

julia> a_schedule = schedule_within([triple, triple], 44100Hz);

julia> extrema(a_schedule) .≈ (-1.0, 1.0)
(true, true)

AudioSchedules.segments — Method

segments(start_level, shape, duration, end_level)

Called by envelope. Return a tuple of pairs in the form (segment, duration), where duration has units of time (like s), with a segment of shape shape. Shapes include Line, Grow, and Hook.

julia> using AudioSchedules

julia> using Unitful: s

julia> segments(1, Grow, 1s, ℯ)
((Grow(1.0, 1.0 s^-1), 1 s),)

AudioSchedules.@q_str — Macro

q"interval"

Create a musical interval. You can specify a numerator (which defaults to 1) and denominator (which defaults to 1) and an octave shift (which defaults to 0).

julia> using AudioSchedules

julia> q"1"
1//1

julia> q"3/2"
3//2

julia> q"2/3o1"
4//3

julia> q"2/3o-1"
1//3

julia> q"o2"
4//1

julia> q"1 + 1"
ERROR: LoadError: Can't parse interval 1 + 1
[...]