Interface

AudioSchedules.AudioSchedule

AudioSchedules.Cycles

AudioSchedules.Envelope

AudioSchedules.InfiniteMap

AudioSchedules.Line

AudioSchedules.Synthesizer

AudioSchedules.make_iterator

AudioSchedules.play

AudioSchedules.restart!

AudioSchedules.schedule!

AudioSchedules.AudioSchedule — Method

AudioSchedule(sink)

Create a AudioSchedule to schedule changes to sink.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> using PortAudio: PortAudioStream

julia> stream = PortAudioStream(samplerate = 44100);

julia> schedule = AudioSchedule(stream.sink)
AudioSchedule with triggers at () seconds

Add a synthesizer to the schedule with schedule!. You can schedule for a duration in seconds, or use an Envelope.

julia> envelope = Envelope((0, 0.25, 0), (0.05s, 0.95s), (Line, Line));

julia> schedule!(schedule, InfiniteMap(sin, Cycles(440Hz)), 0s, envelope)

julia> schedule!(schedule, InfiniteMap(sin, Cycles(440Hz)), 1s, envelope)

julia> schedule!(schedule, InfiniteMap(sin, Cycles(550Hz)), 1s, envelope)

julia> schedule
AudioSchedule with triggers at (0.0, 1.0, 2.0, 3.0, 4.0) seconds

Then, you can play it with play.

julia> play(schedule)

You can only play a schedule once. If you would like to play it again, you must explicitly restart! it.

julia> play(schedule)
ERROR: EOFError: read end of file
[...]

julia> restart!(schedule)

julia> play(schedule)

julia> close(stream)

AudioSchedules.Cycles — Type

Cycles(frequency)

Cycles from 0 2π to repeat at a frequency.

AudioSchedules.Envelope — Type

Envelope(levels, durations, shapes)

Shapes are all functions which return Synthesizers:

shape(start_value, end_value, duration) -> Synthesizer

durations and levels list the time and level of the boundaries of segments of the envelope. For example,

Envelope([0.0, 1.0, 1.0, 0.0], [.05 s, 0.9 s, 0.05 s], [Line, Line, Line])

will create an envelope with three segments:

Line(0.0, 1.0, 0.05 s)
Line(1.0, 1.0, 0.9 s)
Line(s1.0, 0.0, 0.05 s)

See the example for AudioSchedule.

AudioSchedules.InfiniteMap — Type

InfiniteMap(a_function, synthesizers...)

Map a_function over synthesizers, assuming that none of the synthesizers will end early.

AudioSchedules.Line — Type

Line(start_value, end_value, duration)

A line from start_value to end_value that lasts for duration.

AudioSchedules.Synthesizer — Type

abstract type Synthesizer

Synthesizers need only support make_iterator.

AudioSchedules.make_iterator — Method

make_iterator(synthesizer, samplerate)

Return an iterator that will the synthesizer at a given samplerate

AudioSchedules.play — Method

play(schedule::AudioSchedule)

Play an AudioSchedule. See the example for AudioSchedule.

AudioSchedules.restart! — Method

restart!(schedule::AudioSchedule)

Restart a schedule.

AudioSchedules.schedule! — Method

schedule!(schedule::AudioSchedule, synthesizer, start_time, duration)

Schedule an audio synthesizer to be added to the schedule, starting at start_time and lasting for duration. You can also pass an Envelope as a duration. See the example for AudioSchedule. Note: the schedule will discard the first sample in the iterator during scheduling.