Interface

Performance note

Presumably due to the limits of inference, scheduling 16 or more synthesizers simultaneously will lead you off a performance cliff. Hopefully this limitation will go away in future versions of Julia.

AudioSchedules.AudioSchedule

AudioSchedules.Cycles

AudioSchedules.Grow

AudioSchedules.Hook

AudioSchedules.Line

AudioSchedules.Map

AudioSchedules.Plan

AudioSchedules.SawTooth

AudioSchedules.Scale

AudioSchedules.add!

AudioSchedules.add!

AudioSchedules.equal_loudness

AudioSchedules.get_duration

AudioSchedules.make_iterator

AudioSchedules.preview

AudioSchedules.seek_peak

AudioSchedules.segments

AudioSchedules.skip

AudioSchedules.@q_str

AudioSchedules.AudioSchedule — Method

AudioSchedule(plan::Plan)

Return a SampledSource from a Plan.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> plan = Plan(44100Hz)
Plan with triggers at ()

julia> add!(plan, Map(sin, Cycles(440Hz)), 0s, 0, Line => 1s, 1, Line => 1s, 0)

julia> a_schedule = AudioSchedule(plan)
AudioSchedule

You can find the number of samples in an AudioSchedule with length.

julia> the_length = length(a_schedule)
88200

You can use the schedule as a source for samples.

julia> read(a_schedule, the_length)
88200-frame, 1-channel SampleBuf{Float64, 2}
2.0s sampled at 44100.0Hz
▄▄▅▅▅▆▆▆▆▆▆▆▆▆▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▆▆▆▆▆▆▆▆▆▅▅▅▄▄

AudioSchedules.Cycles — Type

Cycles(frequency)

Cycles from 0 to 2π to repeat at a frequency (with frequency units, like Hz). Supports make_iterator.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> first(make_iterator(Cycles(440Hz), 44100Hz))
0.0

AudioSchedules.Grow — Type

Grow(start_level, rate)

Exponentially grow or decay from start_level (unitless), at a continuous rate (with units per time like 1/s). Supports make_iterator and segments.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> first(make_iterator(Grow(1, 1 / s), 44100Hz))
1.0

AudioSchedules.Hook — Type

Hook(rate, slope)

Make a hook shape, with an exponential curve growing at a continuous rate (with units per time like 1/s), followed by a line with slope (with units per time like 1/s). Use with add!. Supports segments. Not all hooks are solvable.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> plan = Plan(44100Hz)
Plan with triggers at ()

julia> add!(plan, Map(sin, Cycles(440Hz)), 0s, 1, Hook(1 / s, 1 / s) => 2s, ℯ + 1)

julia> add!(plan, Map(sin, Cycles(440Hz)), 0s, 1, Hook(1 / s, 1 / s) => 2s, 0)
ERROR: Unsolvable hook
[...]

AudioSchedules.Line — Type

Line(start_level, slope)

A line from start_level (unitless) with slope (with units per time like 1/s). Supports make_iterator and segments.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> first(make_iterator(Line(0, 1 / s), 44100Hz))
0.0

AudioSchedules.Map — Type

Map(a_function, synthesizers...)

Map a_function over synthesizers. Supports make_iterator.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> first(make_iterator(Map(sin, Cycles(440Hz)), 44100Hz))
0.0

AudioSchedules.Plan — Type

Plan(sample_rate)

Create an empty plan for an audio schedule. Use add! to add new synthesizers for plan. Specify a sample_rate with units per time, like 1/s.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> Plan(44100Hz)
Plan with triggers at ()

AudioSchedules.SawTooth — Method

SawTooth(overtones)

Build a saw-tooth wave from its partials, starting with the fundamental (1), up to overtones.

To increase richness but also buziness, increase overtones.

julia> using AudioSchedules

julia> SawTooth(3)(π / 4)
1.4428090415820634

AudioSchedules.Scale — Type

function Scale(ratio)

A simple wrapper that will multiply inputs by the ratio.

julia> using AudioSchedules

julia> Scale(3)(2)
6

AudioSchedules.add! — Method

add!(plan::Plan, synthesizer, start_time)

Add a synthesizer to the plan, where synthesizer must support make_iterator and get_duration.

julia> using AudioSchedules

julia> using FileIO: load

julia> using LibSndFile: LibSndFile

julia> using Unitful: Hz, s

julia> cd(joinpath(pkgdir(AudioSchedules), "test"))

julia> plan = Plan(44100Hz)
Plan with triggers at ()

julia> add!(plan, load("clunk.wav"), 0s)

julia> plan
Plan with triggers at (0.0 s, 0.351859410430839 s)

julia> a_schedule = AudioSchedule(plan)
AudioSchedule

julia> read(a_schedule, length(a_schedule))
15517-frame, 1-channel SampleBuf{Float64, 2}
0.351859410430839s sampled at 44100.0Hz
▇▇▇▇▇▇▇▆▆▆▆▅▅▅▅▅▅▅▅▅▅▅▅▅▅▄▅▅▄▄▅▄▄▄▄▄▄▄▄▄▄▄▃▄▄▄▃▄▄▃▄▄▄▃▃▃▃▃▃▃▃▃▃▃▃▃▃▃▃▃▂▃▃▃▂▃▃▂▂▂

AudioSchedules.add! — Method

add!(plan::Plan, synthesizer, start_time,
    start_level, shape => duration, end_level, more_segments...
)

Add a synthesizer to a Plan, where synthesizer is anything that supports make_iterator, start_time has units of time (like s), and rest of the arguments specify the shape of the envelope.

For all envelope segments, call

segments(shape, start_level, duration, end_level)

duration should have units of time (like s). For example,

add!(plan, synthesizer, start_time, 0, Line => 1s, 1, Line => 1s, 0)

will call segments twice:

segments(Line, 0, 1s, 1)
segments(Line, 1, 1s, 0)

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> plan = Plan(44100Hz)
Plan with triggers at ()

julia> add!(plan, Map(sin, Cycles(440Hz)), 0s, 0, Line => 1s, 1, Line => 1s, 0)

julia> plan
Plan with triggers at (0.0 s, 1.0 s, 2.0 s)

julia> AudioSchedule(Plan(44100Hz))
ERROR: ArgumentError: AudioSchedules require at least one synthesizer
[...]

AudioSchedules.equal_loudness — Method

equal_loudness(synthesizer::Map{<:Any, Tuple{Cycles}})

Change the volume of a synthesizer so that sounds played at different frequencies will have the same perceived volume. Assumes that the map function has a period of 2π.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> soft = equal_loudness(Map(cos, Cycles(10000Hz)));

julia> first(make_iterator(soft, 44100Hz)) ≈ 0.0053035474
true

Technical details: uses the ISO 226:2003 curve for 40 phons. Scales output by a ratio of the equivalent sound pressure at the current frequency to the equivalent sound pressure at 20Hz (about as low as humans can hear).

AudioSchedules.get_duration — Method

get_duration(synthesizer)

Get the duration of a synthesizer (with units of time, like s), for synthesizers with an inherent length.

julia> using AudioSchedules

julia> using FileIO: load

julia> using LibSndFile: LibSndFile

julia> cd(joinpath(pkgdir(AudioSchedules), "test"))

julia> get_duration(load("clunk.wav"))
0.351859410430839 s

AudioSchedules.make_iterator — Method

make_iterator(synthesizer, sample_rate)

Return an iterator that will the play the synthesizer at sample_rate (with frequency units, like Hz). The iterator should yield ratios between -1 and 1. Assumes that iterators will never end while they are scheduled. In addition to supporting iterate, iterators should also support AudioSchedules.preview and AudioSchedules.skip, and iteration should have no side effects.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> using FileIO: load

julia> using LibSndFile: LibSndFile

julia> cd(joinpath(pkgdir(AudioSchedules), "test"))

julia> first(make_iterator(load("clunk.wav"), 44100Hz))
0.00168Q0f15

julia> make_iterator(load("clunk.wav"), 48000Hz)
ERROR: ArgumentError: Sample rate mismatch
[...]

AudioSchedules.preview — Method

preview(iterator, state, ahead)

ahead = 1 is equivalent to getting the item from iterate(iterator, state). Increasing ahead by 1 will be as if you iterated and then discarded.

AudioSchedules.seek_peak — Method

seek_peek(a_schedule::AudioSchedule)

Find the maximum absolute amplitude in an AudioSchedule.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> plan = Plan(44100Hz)
Plan with triggers at ()

julia> add!(plan, Map(sin, Cycles(440Hz)), 0s, 0, Line => 1s, 1, Line => 1s, 0)

julia> seek_peak(AudioSchedule(plan))
0.9994267666261519

AudioSchedules.segments — Method

segments(shape, start_level, duration, end_level)

Called for each envelope segment passed to add!. Return a tuple of pairs in the form (segment, duration), where duration has units of time (like s), for a segment of shape shape.

julia> using AudioSchedules

julia> using Unitful: s

julia> segments(Grow, 1, 1s, ℯ)
((Grow(1.0, 1.0 s^-1), 1 s),)

AudioSchedules.skip — Method

skip(iterator, state, ahead)

ahead = 1 is equivalent to getting the new state from iterate(iterator, state). Increasing ahead by 1 will be as if you iterated and then discarded.

Base.map — Method

map(a_function, schedule::AudioSchedule)

Map a function over all of the synthesizers in the schedule.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> plan = Plan(44100Hz)
Plan with triggers at ()

julia> add!(plan, Map(sin, Cycles(440Hz)), 0s, 0, Line => 1s, 1, Line => 1s, 0)

julia> a_schedule = AudioSchedule(plan)
AudioSchedule

julia> seek_peak(map(Scale(2), a_schedule))
1.9988535332523039

AudioSchedules.@q_str — Macro

q"interval"

Create a musical interval. You can specify a numerator (which defaults to 1) and denominator (which defaults to 1) and an octave shift (which defaults to 0).

julia> using AudioSchedules

julia> q"1"
1//1

julia> q"3/2"
3//2

julia> q"2/3o1"
4//3

julia> q"2/3o-1"
1//3

julia> q"o2"
4//1

julia> q"1 + 1"
ERROR: LoadError: Base.Meta.ParseError("Can't parse interval 1 + 1")
[...]