Interface

AudioSchedules.AudioSchedule

AudioSchedules.Cycles

AudioSchedules.Grow

AudioSchedules.Hook

AudioSchedules.Line

AudioSchedules.Map

AudioSchedules.compound_wave

AudioSchedules.envelope

AudioSchedules.equal_loudness

AudioSchedules.extrema!

AudioSchedules.get_duration

AudioSchedules.make_iterator

AudioSchedules.schedule_within

AudioSchedules.segments

AudioSchedules.@q_str

AudioSchedules.AudioSchedule — Method

AudioSchedule(triples, the_sample_rate)

Return a SampledSource. triples should be a vector of triples in the form

(synthesizer, start_time, duration_or_envelope)

where synthesizer is anything that supports make_iterator, start_time has units of time (like s), and duration_or_envelope is either a duration (with units of time, like s) or an envelope.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> an_envelope = envelope(0, Line => 1s, 1, Line => 1s, 0);

julia> triple = (Map(sin, Cycles(440Hz)), 0s, an_envelope);

julia> a_schedule = AudioSchedule([triple], 44100Hz);

You can find the number of samples in an AudioSchedule with length.

julia> the_length = length(a_schedule)
88200

You can use the schedule as a source for samples.

julia> read(a_schedule, the_length);

The schedule must have at least one triple.

julia> AudioSchedule([], 44100Hz)
ERROR: AudioSchedules require at least one triple
[...]

AudioSchedules.Cycles — Type

Cycles(frequency)

Cycles from 0 to 2π to repeat at a frequency (with frequency units, like Hz). Supports make_iterator.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> first(make_iterator(Cycles(440Hz), 44100Hz))
0.0

AudioSchedules.Grow — Type

Grow(start, rate)

Exponentially grow or decay from start (unitless), at a continuous rate (with units per time like 1/s). Supports make_iterator.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> first(make_iterator(Grow(1, 1/s), 44100Hz))
1.0

AudioSchedules.Hook — Type

Hook(rate, slope)

Make a hook shape, with an exponential curve growing at a continuous rate (with units per time like 1/s), followed by a line with slope (with units per time like 1/s). Use with envelope.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> envelope(1, Hook(1/s, 1/s) => 2s, ℯ + 1)
((Grow(1.0, 1.0 s^-1), 1.0 s), (Line(2.718281828459045, 1.0 s^-1), 1.0 s))

AudioSchedules.Line — Type

Line(start, slope)

A line from start (unitless) with slope (with units per time like 1/s). Supports make_iterator.

julia> using AudioSchedules

julia> using Unitful: Hz, s

julia> first(make_iterator(Line(0, 1/s), 44100Hz))
0.0

AudioSchedules.Map — Type

Map(a_function, synthesizers...)

Map a_function over synthesizers. Supports make_iterator.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> first(make_iterator(Map(sin, Cycles(440Hz)), 44100Hz))
0.0

AudioSchedules.compound_wave — Method

compound_wave(overtones, dampen)

Create a compound wave. Create a function of this form:

radians -> sum((sin(overtone * radians) / overtone for overtone in 1:overtones))

To increase richness but also buziness, increase overtones. To decrease buziness but also buziness, decrease dampen.

julia> using AudioSchedules

julia> compound_wave(3)(π/4)
1.4428090415820634

AudioSchedules.envelope — Method

envelope(start_level, shape => duration, end_level, more_segments...)

For each envelope segment, call

segments(start_level, shape, duration, end_level)

duration should have units of time (like s). For example,

envelope(0, Line => 1s, 1, Line => 1s, 0)

will add two segments:

segments(0, Line, 1s, 1)
segments(1, Line, 1s, 0)

julia> using AudioSchedules

julia> using Unitful: s

julia> envelope(0, Line => 1s, 1, Line => 1s, 0)
((Line(0.0, 1.0 s^-1), 1 s), (Line(1.0, -1.0 s^-1), 1 s))

AudioSchedules.equal_loudness — Method

equal_loudness(synthesizer::Map{<:Any, Tuple{Cycles}})

Change the volume of a synthesizer so that sounds played at different frequencies will have the same perceived volume. Assumes that the map function has a period of 2π.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> soft = equal_loudness(Map(cos, Cycles(10000Hz)));

julia> first(make_iterator(soft, 44100Hz)) ≈ 0.0053035474
true

Technical details: uses the ISO 226:2003 curve for 40 phons. Scales output by a ratio of the equivalent sound pressure at the current frequency to the equivalent sound pressure at 20Hz (about as low as humans can hear).

AudioSchedules.extrema! — Method

extrema!(a_schedule::AudioSchedule)

Find the extrema of a_schedule. This will consume the schedule.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> triple = (Map(sin, Cycles(440Hz)), 0s, 1s);

julia> extrema!(AudioSchedule([triple, triple], 44100Hz)) .≈ (-1.9999995, 1.9999995)
(true, true)

AudioSchedules.get_duration — Method

get_duration(synthesizer)

Get the duration of a synthesizer (with units of time, like s), for synthesizers with an inherent length.

julia> using AudioSchedules

julia> using FileIO: load

julia> import LibSndFile

julia> cd(joinpath(pkgdir(AudioSchedules), "test"))

julia> get_duration(load("clunk.wav"))
0.3518820861678005 s

AudioSchedules.make_iterator — Method

make_iterator(synthesizer, the_sample_rate)

Return an iterator that will the play the synthesizer at the_sample_rate (with frequency units, like Hz). The iterator should yield ratios between -1 and 1. Iterators will typically be infinite, with endings determined by the schedule instead.

julia> using AudioSchedules

julia> using Unitful: Hz

julia> using FileIO: load

julia> import LibSndFile

julia> cd(joinpath(pkgdir(AudioSchedules), "test"))

julia> first(make_iterator(load("clunk.wav"), 44100Hz))
0.00168Q0f15

AudioSchedules.schedule_within — Method

schedule_within(triples, the_sample_rate; maximum_volume = 1.0)

Make an AudioSchedule with triples and the_sample_rate, then adjust the volume to maximum_volume.

julia> using AudioSchedules

julia> using Unitful: s, Hz

julia> triple = (Map(sin, Cycles(440Hz)), 0s, 1s);

julia> a_schedule = schedule_within([triple, triple], 44100Hz);

julia> extrema!(a_schedule) .≈ (-1.0, 1.0)
(true, true)

AudioSchedules.segments — Method

segments(an_envelope, shape, duration, start_level, end_level)

Called by envelope. Return a tuple of pairs in the form `(segment, duration), where duration has units of time (like s).

julia> using AudioSchedules

julia> using Unitful: s

julia> segments(1, Grow, 1s, ℯ)
((Grow(1.0, 1.0 s^-1), 1 s),)

AudioSchedules.@q_str — Macro

q"interval"

Create a musical interval. You can specify a numerator (which defaults to 1) and denominator (which defaults to 1) and an octave shift (which defaults to 0).

julia> using AudioSchedules

julia> q"1"
1//1

julia> q"3/2"
3//2

julia> q"2/3o1"
4//3

julia> q"2/3o-1"
1//3

julia> q"o2"
4//1

julia> q"1 + 1"
ERROR: LoadError: Can't parse interval 1 + 1
[...]