Interface

AudioSchedules.AudioSchedule

AudioSchedules.Cycles

AudioSchedules.Envelope

AudioSchedules.Line

AudioSchedules.Plan

AudioSchedules.StrictMap

AudioSchedules.Synthesizer

AudioSchedules.make_iterator

AudioSchedules.schedule!

AudioSchedules.AudioSchedule — Method

AudioSchedule(the_sample_rate)

Create an AudioSchedule.

julia> using AudioSchedules

julia> using SampledSignals: s, Hz

julia> a_schedule = AudioSchedule(44100Hz)
AudioSchedule with triggers at () seconds

Add a synthesizer to the schedule with schedule!.

julia> envelope = Envelope((0, 0.25, 0), (0.05s, 0.95s), (Line, Line));

julia> schedule!(a_schedule, StrictMap(sin, Cycles(440Hz)), 0s, envelope)

julia> schedule!(a_schedule, StrictMap(sin, Cycles(440Hz)), 1s, envelope)

julia> schedule!(a_schedule, StrictMap(sin, Cycles(550Hz)), 1s, envelope)

julia> a_schedule
AudioSchedule with triggers at (0.0, 0.05, 1.0, 1.05, 2.0) seconds

Then, you can create a SampledSource from the schedule using Plan.

julia> using SampledSignals: unsafe_read!

julia> a_plan = Plan(a_schedule);

You can find the number of samples in a Plan with length.

julia> the_length = length(a_plan)
88200

You can use Plan as a source for samples.

julia> buf = Vector{Float64}(undef, the_length);

julia> unsafe_read!(a_plan, buf, 0, the_length);

julia> buf[1:4] == [0.0, 1.417806391094423e-5, 4.23948845413273e-5, 8.44006285522918e-5]
true

AudioSchedules.Cycles — Type

Cycles(frequency)

Cycles from 0 2π to repeat at a frequency.

AudioSchedules.Envelope — Type

Envelope(levels, durations, shapes)

Shapes are all functions which return Synthesizers:

shape(start_value, end_value, duration) -> Synthesizer

durations and levels list the time and level of the boundaries of segments of the envelope. For example,

Envelope([0.0, 1.0, 1.0, 0.0], [.05 s, 0.9 s, 0.05 s], [Line, Line, Line])

will create an envelope with three segments:

Line(0.0, 1.0, 0.05 s)
Line(1.0, 1.0, 0.9 s)
Line(1.0, 0.0, 0.05 s)

See the example for AudioSchedule.

AudioSchedules.Line — Type

Line(start_value, end_value, duration)

A line from start_value to end_value that lasts for duration.

AudioSchedules.Plan — Method

Plan(a_schedule::AudioSchedule)

Return a SampledSource for the schedule.

AudioSchedules.StrictMap — Type

StrictMap(a_function, synthesizers...)

Map a_function over synthesizers, assuming that none of the synthesizers will end before they are scheduled to.

AudioSchedules.Synthesizer — Type

abstract type Synthesizer

Synthesizers need only support make_iterator.

AudioSchedules.make_iterator — Method

make_iterator(synthesizer, the_sample_rate)

Return an iterator that will the play the synthesizer at the_sample_rate

AudioSchedules.schedule! — Method

schedule!(schedule::AudioSchedule, synthesizer::Synthesizer, start_time, envelope::Envelope)

Schedule an audio synthesizer to be added to the schedule, starting at start_time with the duration and volume contained in an Envelope. See the example for AudioSchedule.